+++ db/schema.rb
@@ -293,19 +293,19 @@ ActiveRecord::Schema.define(version: 20140923185707) do
     t.string   "image"
     t.integer  "sms_code"
     t.integer  "confirm_code"
-    t.decimal  "credit",                 default: 10.0
+    t.integer  "credit",                 default: 10
     t.datetime "last_seen_at"
     t.datetime "last_visited_inbox_at"
     t.integer  "num_following",          default: 0
     t.integer  "num_followers",          default: 0
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.string   "email",                  default: "",   null: false
-    t.string   "encrypted_password",     default: "",   null: false
+    t.string   "email",                  default: "", null: false
+    t.string   "encrypted_password",     default: "", null: false
     t.string   "reset_password_token"
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
-    t.integer  "sign_in_count",          default: 0,    null: false
+    t.integer  "sign_in_count",          default: 0,  null: false
     t.datetime "current_sign_in_at"
     t.datetime "last_sign_in_at"
     t.string   "current_sign_in_ip"
